refactor: rename from *-dto.ts to *.dto.ts and *-template.ts to *.template.ts
