+++ src/auth/auth.controller.ts
@@ -16,9 +16,9 @@ import { RefreshAuthGuard } from './guards/refresh-auth.guard'
 import { Request, Response } from 'express'
 import { User } from '@prisma/client'
 import { COOKIE_CONFIG } from './constant'
-import { RegisterDTO } from './dto/register-dto'
-import { VerifyRegistrationDTO } from './dto/verify-registration-dto'
-import { CreateUserDTO } from './dto/create-user-dto'
+import { RegisterDTO } from './dto/register.dto'
+import { VerifyRegistrationDTO } from './dto/verify-registration.dto'
+import { CreateUserDTO } from './dto/create-user.dto'
 import { PreRegistGuard } from './guards/pre-regist.guard'
 import { GetUser } from 'src/common/decorators/getUser.decorator'
