+++ src/auth/auth.service.ts
@@ -9,11 +9,11 @@ import { EmailService } from 'src/email/email.service'
 import { PrismaService } from 'src/prisma/prisma.service'
 import { customAlphabet } from 'nanoid'
 import { Prisma, User } from '@prisma/client'
-import { CreateUserDTO } from './dto/create-user-dto'
-import { verificationCodeTemplate } from './templates/verification-code-template'
-import { VerifyRegistrationDTO } from './dto/verify-registration-dto'
-import { RegisterDTO } from './dto/register-dto'
-import { newUserRegistrationTemplate } from './templates/new-user-registration-template'
+import { CreateUserDTO } from './dto/create-user.dto'
+import { verificationCodeTemplate } from './templates/verification-code.template'
+import { VerifyRegistrationDTO } from './dto/verify-registration.dto'
+import { RegisterDTO } from './dto/register.dto'
+import { newUserRegistrationTemplate } from './templates/new-user-registration.template'
 import * as bcrypt from 'bcryptjs'
 import { JwtService } from '@nestjs/jwt'
 import { Cron, CronExpression } from '@nestjs/schedule'
