+++ src/auth/templates/new-user-registration.template.ts
